+++ apps/tests/group-docs/realtime.config.ts
@@ -28,6 +28,7 @@ export const {
   useSubscribeMe,
   useBroadcast,
   useBroadcastListener,
+  useDuplicate,
 } = createRealtimeGroupContext<PresenceData>()
 
 export const {