+++ packages/@inrealtime/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inrealtime/react
 
+## 0.2.9
+
+### Patch Changes
+
+- Added duplicate functionality for group documents"
+
 ## 0.2.8
 
 ### Patch Changes